deb: Add stand alone mode to changelogger
[why]
The version number of untagged HEADs is determined by Meson (because it
is so 'complicated' and we want to make sure a correct version even if
the git log is corrupt) and filled in by text substitution.

This becomes impossible if we want to use it as full stand alone script
that can be executed from the command line (i.e. ttf-scripts package).

[how]
Determine in-between version numbers with patch numbers if a package
name is specified on the command line.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>