Add SmallVector container
[why]
The SmallVector class is basically a vector which can
store a limited number of elements in the object itself,
similar to the small-string optimization in std::string.
This can be useful to avoid heap allocations.

This container should be very handy as an alternative
return type for functions such as split() which now
return a std::vector. This means that these functions
always have to allocate, which is slow and also requires
synchronization in multithreaded environments. A
SmallVector can fit entirely on the stack, while still
retaining the flexibility to go to the heap if more
storage is required.

[how]
SmallVector has been written from scratch with the
specification of std::vector as a blueprint. Where
possible, it behaves in an identical or at least
similar manner.