tokenize: Allow associative containers as target
[why]
There is no need that our emplace helper can not work with more standard
library containers.

[how]
Detect if the container has an emplace_back() function. If it has use it
for emplacing. If not, use emplace() that might or might not exist and
fit the signature.

This enables all the set variants.

All maps are still unsupported because, well they need two things to
work in any meaningful manner. So although technically possible to
support them out of the box I see no useful way to do so.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>