changelogger: Put version extraction into function
[why]
We extract the version number/string in three places with identical sed
calls. While this is the most crucial part of the automatic version
determination it is almost undocumented - and spread around in the code.

[how]
Use one function to extract the version number/string. This avoids
having different code in the places where it is used and a but of
documentation can be added now.

Reported-by: Soeren Grunewald <soeren.grunewald@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>