Disable debian packaging if we have a git problem
[why]
When we do not have git or we have git but our source is not a git repo
but some tarball we fail to set up 'doocs-release'.

[how]
A doocs-release on Linux automatically also includes setup of debian
packaging. For the packaging we need the git tags to extract the actual
version we want to pack and autocreate a changelog from the commits.

When the topmost git tag can not be determined (for whatever reason)
there is something wrong with the git setup/readout/repo, and we refuse
to prepare debian packaging.

Fixes: #5

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>