tokenize: Always initialize token container
[why]
We can specify any container that holds the tokens as template
parameter. When the container is a class it will be automatically
initialized.
But if we choose to have a basic type as container, that is not
initialized and tokenize() is not usable (in an easy way).

[how]
Just specify that the container is to be initialized. This will not
change the generated code on class type containers, but for containers
of basic type.

[note]
Here is an example, where tokenize becomes usable with this commit:

// Check for presence of a specific token in a token list
auto to_find = "4";
auto is_present = gul14::tokenize<bool>(
    "10  3 5  12 ", gul14::default_whitespace_characters,
    [&to_find] (bool& present, gul14::string_view s) {
        present |= (s == to_find);
    });

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>