string_util: Fix static_assert() on GCC 5.x
[why]
GCC 5 on Ubuntu 16.04 has problems with the constexpr-ness of the std::array
size that we would like to check.
Clang on Ubuntu 16.04 runs with no problems.

[how]
Just do not assert the array size on that old machine.

Alternatively one could access the array_traits (that we are not allowed
to access) for the test. This guaranteed to succeed on Ubuntu 16.04 but
can fail on other or even older platforms:

 inline std::string hex_string(Integer v)
 {
-    static_assert(hex_digits.size() == 16, "hex_digits has wrong size");
+    #if (defined(__GNUC__) && !defined(__clang__) && __GNUC__ < 6)
+        // Old GCC on Ubuntu 16.04 has constexpr problems, so we access the
+        // array traits
+        static_assert(sizeof(decltype(hex_digits)::_M_elems) == 16, "hex_digits has wrong size");
+    #else
+        static_assert(hex_digits.size() == 16, "hex_digits has wrong size");
+    #endif

Fixes: #6

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>