string_util: Correct vimline
[why]
It instructs the vi to embed tabs into the source. This is exactly what
we do not want.

[how]
Use the same vimline we have in other source files.
Well, they are not unified ...

This drops the type instruction (which all vis should manage anyhow).
This drops the line length instruction (we do not have it elsewhere).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>