ci: Raise parallelism
[why]
We can work on multiple platforms at once.

[how]
Start building on 'other' platforms (all in parallel) when the
first platform compiled successfully.

Do not run the compilation on all platforms in parallel right away,
because that would create all the jobs even if it is hopeless
(because there is a syntax error or something, that will cause all
platforms to fail).

Start tests and afterwards install steps for each platform when it has
compiled, regardless of other platforms.

Benefit is that one can see that it succeeds on for ex MacOS, although
it fails on Xenial which would be processed earlier in the normal
pipeline.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>