to_number: Fix handling of high resultion 'long double' types
[why]
On aarch64 the mantissa of long double is bigger then 64 bit and does
not fit into our mantissa holding integer variable.

While the type's sizeof() is 128 bit on most platforms we have, the actual
bits a long double occupies is only 80 on x64. The part that is the
mantissa fits easily into 64 bit on that platforms.

[how]
To change not to much code we just use a 128 bit integer in that cases.

First FloatConversionIntType is changed into a template that depends
on the floating point number's type. So we can use different
bitness for the integer mantissa representation.

If the mantissa bitness of long double on a platform does not fit into
uint64_t and that platform supports __uint128_t we use that (only for
long double).

If the platform does not support __uint128_t (MSVC) but 64 bits are not
enough a static_assert will raise an error.

__uint128_t is a non standard extension present in GCC, clang, and intel.

If __uint128_t is available the compiler sets __SIZEOF_INT128__:
https://gcc.gnu.org/legacy-ml/gcc-help/2015-08/msg00185.html
Tested on GCC, clang, and intel, some compiler versions each, on Godbolt.

Fixes: #7

Reported-by: Gerhard Schlesselmann <gerhard.schlesselmann@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>