ci: Add Windows 64 test
[why]
We want to check Windows builds...

[how]
I have a Windows box that always runs and that I use for PLC debugging
occasionally (like once every couple of months).
We can use that to host a runner.

If it happens to be 'off' (like after a power glitch or so) the steps
are optional for GUL pipeline success. So it is more like
'informational'.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>