meson: Allow non-packaging build with mismatched verions
[why]
If there is a mismatch between git tag version and API version compiling
is not possible.
In some cases one wants to compile the library and does not care about
the version number...

[how]
Important might be to have the correct verison number when the library is
packaged.

So we error out on a debian packaging prepared build, but only issue
a message when not. This can be dangerous when the packaging is desired
but done to non-debian packages, or if the build is intended to be
installed directly into the system.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>