ci: Use templates to simplify definitions
[why]
It is confusing to have so many stages, and absolutely unneeded.
It is error prone to repeat the machine setup for each job on that target.
It is error prone to repeat the meson setup for each target.

[how]
Use templates to extract common features of the jobs into hidden job
templates.
Also adhere to the recommended stage names where possible.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>