ci: Add UB sanitizer
[why]
It seems useful and is a reccomended best practice, to check for
undefined behaviour that is utilized.

[how]
Set a second sanitizer run up. Although we could in principle join the
address and the undefined sanitizer the compilation takes rather long in
the combined case, and we run into variable tracking limitations with
the already complex functions generated by Catch2.

[note]
Here is a list of the checks done:
https://gcc.gnu.org/onlinedocs/gcc-11.2.0/gcc/Instrumentation-Options.html

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>