ci: Create deb package in gitlab packages repo
[why]
It is easier if we have a way to access the latest successful build as a
debian package from the gitlab package repository.

[how]
If the pipelines processes the master branch we publish the created
debian packages in the gitlab package repo.

Also fix small issue with the debian/rules: The $PWD envvar is not set
in the non-interactive containers in gitlab. use `pwd` instead.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>