ci: Fix 'Simplify CI setup for xenial and focal'
[why]
When we try to upload the package we use 'curl' but that has not
been installed yet, so builds with deploying into the package
repository will fail (i.e. pipelines on master branch).

[how]
Install the curl package.

[note]
We install some packages for all steps and some just for the
deploy step. That can be unexpected and I would rather change
that to install all possibly needed packages at once.
Otoh, maybe the current behavior is better, as no unneeded work
is done.
I don't know.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>