ci: Package release on tagging
[why]
The packaging is not triggered if just a tag is added. But that is
needed if nothing with GUL changes but the DOOCS version increases from
time to time (via tag).
Originally we always had some commit to GUL on new DOOCS version, iirc.

Also it seems the packaging does not work on pushes either.

[how]
Reformulate rules to trigger on normal merge request merges.

Add rule to trigger packaging on new tags, if the tag is on the master
branch. 'master' is hard coded and not flexible as $CI_DEFAULT_BRANCH or
similar.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>