type_name: Correct test for Apple clang 13
[why]
On out Darwin buildhost one test for type_name fails.

The reason seem to be that - different from all other compilers before -
the std::string seems to be a self-standing type and not a template
(specialisation of std::basic_string). Or at least is reported as such.

On all other compilers a 'std::string' is reported as
'std::basic_string<char>', but here we get the more used name.

Which is quite good I think, but still different.

[how]
Allow both "basic_string" and "std::string" be returned as the type name
of a string in the test.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>