Create releases with debug symbols
[why]
It might come in handy to have debug symbols for an installed library.

[how]
Use 'debugoptimized' rather than 'release' builds.

Debian will package those libraries stripped and the debug symbols as
extra package, so the impact here is just that we have only -O2 instead
of -O3.

Well, and -Ddebug will be set, which means asserts will be evaluated in
the final lib.