statistics: Allow accumulate() to run on non-fundamental types
[why]
Because the functions in statistics (that use gul14::accumulate) work
always on fundamental types, and for that types pass-by-value is often
better then pass-by-const-ref, the default ElementAccessor<>() has been
implemented to utilize pass-by-value, and been restricted to fundamental
types.

This is fine for the use cases of accumulate() that statistics itself
uses. But if gul::accumulate() should be used in a broader context to
replace std::accumulate(), it needs to support non-fundamental types
also; and use pass-by-const-ref for those.

At the moment that is not possible out of the box and one needs to
supply a user defined accessor function. That makes use of
gul::accumulate() in that instances complex and error prone. It would be
better if we would supply a more generic default accessor (one
that covers more use cases).

Pseudo code:

  vec1 = std::vector<int>{ ... };
  struct ut { int v; };
  vec2 = std::vector<ut>{ ... };

currently possible:

  auto sum = gul14::accumulate(vec1, std::plus<int>);
  // equal to
  auto sum = gul14::accumulate(vec1, [](auto a, auto el) { return a + el; });

not possible:

  auto sum = gul14::accumulate(vec2, [](auto a, auto el) { return a + el.v; });

[how]
Lift the restriction that the ElementAccessor<>() can only be
instantiated for fundamental types. Rather have it instantiated for all
types, but use pass-by-value only for fundamental (and pointer) types,
and use pass-by-const-ref in all other cases.

[note]
Also add some tests. Especially tests that compare std::accumulate() to
gul14::accumulate() as example.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>