statistics: ElementAccessor always pass-by-ref
[why]
Even for the simplest types there is no real benefit of passing the
values by-value; the scope of the variable is so small that the
generated code is almost indistinguishable and just differs if the
dereference is done before or after the accessor returns.

With -O2 (GCC) or -O1 (clang) it is all optimized away.

Removing the different handling simplifies the code.

Additionally some fundamental types (do you hear me, long double?) are
slightly more complex to handle with a by-value parameter call and can
benefit from the by-reference parameter scheme.

[how]
Let the default Accessor return a const ref.
Decay type from the Accessor to get the basic type.
Adapt (now more or less useless) tests for pass-by-value.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>