SlidingBuffer: Declare iterator as class and not struct
[why]
It has been defined as `struct` instead of `class` just because GNU's
standard lib had the std::iterator defined as struct. When a class
inherits a struct that lead to a warning.

Now that we do not use std::iterator anymore we can make this a normal
class. Not that it changes anything, technically.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>