rpm: Add some extras for RHEL7
[Why]
On the compute cluster we still use RHEL7. So we would like to
build it here as well. Unfortunately RHEL7 comes with a pretty
old compiler, so we have to make use of Software Collections.

Anyhow, there is a bug in the current meson package (0.55.1)
that in combination with rpm 4.11 prevents us from building an
rpm.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>