tests: Fix compiler warning in statistics test
[why]
Some compiler warnings about narrowing in MSVC.

[how]
According to the gul14::accumulate() documentation the accumulator type
has to be specified in the first template parameter. We did not do that.
It defaults to statistics_result_type which is a floating point type.
Accumulation integers in it might result in precision loss.

So we just give the accumulator type in the template instantiation.

[note]
Also fixed warning about std::vector creation with emplace where the
number is given as int but need to be converted to float.

Fixes: #2 (partially)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>