string_util: Fix compiler warning
[why]
With MSVC there is a warning regarding narrowing...

[how]
In fact we can not shift a long long and store the result in a mere int.
Usually.

Alternatively we could

        unsigned int byte = (u >> (8 * idx)) & 0xff;

and just store the byte like the name implies. But that would require 3
mask operations instead of just 2 which is the lower limit.

Fixes: #2 (partially)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>