Avoid integer overflow by changing calc order
[why]
The lcm was calculated as (a * b) / gcd(a, b).
The product a * b could easily overflow in the
CommonType of a and b.

[how]
Calculate lcm as (a / gcd(a, b)) * b. The division
is guaranteed to give the correct result because
gcd is a divisor of a by definition.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>