Remove "not-later-than" tests for Trigger
[why]
Many Trigger-related tests fail on GitHub MacOS runners
because the systems are under heavy contention. This
sometimes leads to delays on the order of several tens
of milliseconds, which is outside of the tolerance of
the delay tests.

Strictly speaking, there is no guarantee that Trigger's
wait functions wake up after any particular time span at
all, so we should not test for that.

[how]
Remove all tests for Trigger::wait*() waking up no later
than after a certain time span. Adapt the test
descriptions accordingly.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>