Add automatic mirroring to Gitlab
[why]
While development takes place on Github some stuff will still be handled
by the internal Gitlab repository (for example the Doocs release).

Unfortunately the automatic mirroring on Desy's Gitlab instance works
only in `push` and not in `pull` direction. You would have to pay for
that simple switch.

[how]
Use a CI pipeline with a 'cron' trigger to automatically mirror the
repo.

The problem arises that Gitlab does allow only for one workflow (called
pipeline). To overcome that and properly disentangle the two workflows
(i.e. 'build' and 'sync') a new pipeline is created.

That pipeline (mind: Gitlab has only one pipeline) decides if this
trigger shall be a sync or a build trigger. It triggers the secondary
(child) pipeline accoringly.

So:
Put existing pipeline in subdir.
Put sync pipeline in subdir.
Replace existing pipeline with simple one that triggers the right child.

The desicion is taken by the variable `GIT_SYNC`.
Create a Schedule that sets that variable to "1".
Normal pull requests etc will not have set that variable.

Some other prerequisites are needed to push to Gitlab. That is described
in the workflow ... aeh pipeline file.

Pushing is done with `ci:skip` set. Maybe we want that, maybe not.
At the moment we do not detect if there have really been any changes or
not. Room for improvement.