Derive bad_expected_access<T> from bad_expected_access<void>
[why]
The C++23 standard mandates the exception class bad_expected_access<void>
to be the base class for bad_expected_access<T>. This was not the case
in the original version of the implementation.

[how]
Implement bad_expected_access<void> as a class template specialization
of bad_expected_access<T> and derive bad_expected_access<T> from it.
Also, implement the special member functions as protected for
bad_expected_access<void> to avoid slicing (also mandated by the
standard).

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>