Add OverloadSet and make_overload_set()
[why]
Without C++17, we do not have "if constexpr". This can make it hard to
construct a suitable visitor function for gul14::visit(). The
alternative to using "if constexpr" is constructing an overload set of
functions.

[how]
Provide a variadic class template OverloadSet that inherits from an
arbitrary number of function objects (typically lambdas) and imports
their definitions of operator() into an overload set for its own
operator(). Because constructing such an OverloadSet object manually is
almost impossible without C++17's advanced template argument deduction,
we also provide a helper function make_overload_set() to take care of
that.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>