Revert "meson: Resolve warning with newer meson versions"
[Why]
The commit is redundant and even worst, it breaks CI in other projects.
First of using add_global_arguments is a bad idea in general, if at all
we should have used add_project_arguments.
Secondly we assume that every compiler which is not running on windows
does support these flags.
But the most embarrassing thing is, that we already had added the
flags years ago.

Fixes: 866ee07c (meson: Resolve warning with newer meson versions)
Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>
Reported-by: Fini Jastrow <ulf.fini.jastrow@desy.de>