meson: Fix possible issue during package creation
[Why]
With commit 866ee07ca we added checks to run_command, since it
generates warnings with newer meson versions. But this was done in
a hurry without proper thinking, so we totally overlooked the fact
that we check the return code of run_command. So we have to allow
that commands can fail.

Fixes: 866ee07ca (meson: Resolve warning with newer meson versions)
Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>
Reported-by: Fini Jastrow <ulf.fini.jastrow@desy.de>