Meson: Separate build of static and shared library
[why]
Both the shared and the static library were built with a single
both_libraries() command. Unfortunately, the macro
GUL_COMPILING_SHARED_LIB was set in both.

[how]
Separate the build into two commands, a static_library() and a
shared_library(). Define GUL_USING_STATIC_LIB_OR_OBJECTS both
when compiling a static lib and when using it.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>