Introduce TaskHandle and split off ThreadPoolEngine
This commit adds a TaskHandle class template that allows retrieving task
results via a std::future. It also has member functions that allow testing
for completion, running and pending tasks.

To make this safe, we have to split a ThreadPoolEngine off from the main
ThreadPool class. The user-facing class, ThreadPool, creates a
ThreadPoolEngine via make_shared() and stores weak_ptrs to it in the
TaskHandles. This way, each TaskHandle can safely keep the engine alive
for the duration of an inquiry.

As a third major change, we need packaged_tasks to provide futures for
our task return values, but we cannot store a packaged_task inside a
std::function because the former is not copyable. Instead, we now
create a custom type erasure class "NamedTask" for our packaged_task and
also store the task name in it. Both parts are only needed when the task
actually reaches the front of the queue. The queue can now hold simple
unique_ptrs to the named tasks while the ID and the start time remain
direct members of the queue elements.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>