meson: Fix deprecated warning
[why]
Meson 1.3.99 gives these warnings:

NOTICE: Future-deprecated features used:
 * 0.56.0: {'meson.build_root'}
WARNING: Broken features used:
 * 1.3.0: {'str.format: Value other than strings, integers, bools, options, dictionaries and lists thereof.'}

[how]
We try to get the absolute path of a generated file. Since 1.4.0 we have
a get_path() member function on the generated file object, but we can
not use that on older Mesons.

So instead we just manually assemble the filename in the place where the
file is generated and use that later on. Both warned-about functions
removed.

Fixes: #64

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>