string_view: Fix signed mismatch warning
[why]
Apple clang version 15.0.0 warns:

warning: implicit conversion changes signedness:
    'size_type' (aka 'unsigned long') to 'difference_type' (aka 'long')
    [-Wsign-conversion]

[how]
Distances can be forward and backward and thus need a signed number.
Size and index is often described by an unsigned number, to double the
range and to avoid meaningless negative length.

This has historically been mixed without much thought, although in extreme
cases the results were wrong. But usually we do not have strings that long.

To fix this first the private reverse_distance() function is removed, because
it returns a unsigned number and must return a unsigned number from what it
does internally and that does not reflect the concept of 'distance'.

Instead we take the forward distance and do the 'harder' calculations in
the places where reverse_distance() has been used (2 places).

There are now 4 instances of mixing distance and size/index. Upon careful
check it is impossible that the distance in these cases can become
negative; so that we can remove the sign in confidence. This is
expressed by static_casting to the size_type, whereupon the warnings
disappear.

Fixes: #65

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>