time_util: Make time-dependent tests more robust
[why]
One of the tests tries to ascertain that sleep(0) or sleep(-5s) does not
wait. Strictly speaking, that is impossible to test because the program
execution may be interrupted at any time.

[how]
Increase the allowed time duration from 1 ms to 100 ms and cross
fingers. The only other option I see is to remove the test for good.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>