Doxygen: Display "topics" in addition to "modules"
[why]
We used Doxygen's \addtogroup command to define entries for our header
files. In Doxygen 1.8.17 (Ubuntu 20), this category was selected for
display with the "modules" types. Doxygen 1.9.8 has learnt about C++20
modules and reserves the "modules" type for them, so we have to use the
new "topics" keyword.

[how]
Select both "modules" and "topics" for being displayed. Doxygen 1.8.x
ignores the "topics" and Doxygen 1.9.8 does not show anything additional
for "modules" because we do not have any C++ modules in GUL14.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>