ThreadPool: Create threads lazily
[why]
If the resources should be kept low but potentially we need a lot of
parallel workers in some rare circumstances it would be better to
create the threads on demand and not upfront.

[how]
Just do not create any thread on creation but only - if need be and
allowed - on adding a task.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>