ThreadPool: Allow changes in number of threads at runtime
[why]
To dynamically react to load it would be nice to be able to change the
number of threads at runtime.

[how]
To be able to detect finished threads a new struct Thread is used.
It contains the std::thread and a flag that is set to false if the
thread ends itself.

Now if we have too many threads the threads detect it and do not take up
new Tasks but end theirselves.

Ended threads are joined and removed from the vector of Threads.

If the number of threads increased the new threads will be created and
utilized only lazily - when new Tasks are added.

Fixes: #80

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>