to_number: Improve implementation for bool
[Why]
During review some concerns were raised that the basic implementation
would be slow and does not cover some possible use cases.

[How]
Iterate over the string and check if all the characters are digits.
If not, we just check for "true" and "false" case insensitive. If all
characters are numbers we check if the number is different from zero.
So we can do the conversion similar to what C/C++ does.
And of course we add some extra test cases to cover the new
possibilities.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>