to_number: Reduce scope of bool implementation
[Why]
All other to_number implementation respect the actual data range of
there respecting type. E.g. for uint8_t (aka byte) it would be 0..255
everything is threaded as nullopt since it is exceeds the range.
For bool the range is 0..1 so we allow only these values.

Unlike the integer numbers we allow no leading zeros ("00") is not
allowed for 'false'.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>
Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>