Fix: Release Version 2.12.0
[why]
The git-tag based version is 2.12.0 while the API version is still set
to 2.11.

The API version is hardcoded in the /meson.build file, to allow building
from a tarball. The git-tag version is used for packaging purposes, as
the API version major.minor might be the same for bugfix packages.

[how]
Adjust the API version, nut just by setting the git-tag but also in the
Meson specification.

Reported-by: Manuel Mommertz <manuel.mommertz@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>