Meson: Honor value of default_library option
[why]
So far, we always provided build targets for both the static and the
shared library in our build script. This did not allow users to simply
select their desired library type via the standard "default_library"
option.

[how]
Set a project default of "default_library=both" and provide the shared
and static library build targets only if they are part of the
user-selected default_library option.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>